Samples/Freeform-Tool-Calling/Python/requirements.txt
openai
pandas
scikit-learn
matplotlib
python-dotenv